reference clause with both bibitems and reference subclauses: https://github.com/metanorma/metanorma-jis/issues/125
